SASS-1810 : Add Student Loans Journey
