+++ src/test/resources/journeys.conf
@@ -48,6 +48,16 @@ journeys {
       run-if = ["journeyFlows"]
     }
 
+    income-tax-submission-gift-aid-journeys = {
+      description = "Income Tax Submission Journey - Agents/Individuals Gift Aid"
+      load = 1
+      parts = [
+        individual-gift-aid,
+        agent-gift-aid
+      ]
+      run-if = ["journeyFlows"]
+    }
+
     income-tax-submission-employment-journeys = {
       description = "Income Tax Submission Journey - Agents/Individuals Employments"
       load = 1
@@ -64,17 +74,8 @@ journeys {
         agent-add-expenses,
         individual-add-new-employment,
         agent-add-new-employment
-
-      ]
-      run-if = ["journeyFlows"]
-    }
-
-    income-tax-submission-gift-aid-journeys = {
-      description = "Income Tax Submission Journey - Agents/Individuals Gift Aid"
-      load = 1
-      parts = [
-        individual-gift-aid,
-        agent-gift-aid
+        individual-student-loans
+        agent-student-loans
       ]
       run-if = ["journeyFlows"]
     }
