+++ src/test/scala/uk/gov/hmrc/perftests/itsass/IncomeTaxSubmissionSimulation.scala
@@ -17,654 +17,47 @@
 package uk.gov.hmrc.perftests.itsass
 
 import uk.gov.hmrc.performance.simulation.PerformanceTestRunner
-import uk.gov.hmrc.perftests.itsass.AuthLoginAPIRequests.insertAuthRecordInd
-import uk.gov.hmrc.perftests.itsass.AuthLoginRequests._
-import uk.gov.hmrc.perftests.itsass.BenefitsRequests._
-import uk.gov.hmrc.perftests.itsass.DividendsRequests._
-import uk.gov.hmrc.perftests.itsass.EmploymentsRequests._
-import uk.gov.hmrc.perftests.itsass.GiftAidRequests._
-import uk.gov.hmrc.perftests.itsass.IncomeTaxSubmissionRequests._
-import uk.gov.hmrc.perftests.itsass.InterestRequests._
-import uk.gov.hmrc.perftests.itsass.ExpensesRequests._
-import uk.gov.hmrc.perftests.itsass.CrystallisationRequests._
-import uk.gov.hmrc.perftests.itsass.NRSProxyRequests.postNrsProxyStatement
-import uk.gov.hmrc.perftests.itsass.RequestsHelper.{taxYear, taxYearEOY}
+import uk.gov.hmrc.perftests.itsass.simSteps._
 
-class IncomeTaxSubmissionSimulation extends PerformanceTestRunner {
+class IncomeTaxSubmissionSimulation extends PerformanceTestRunner with DividendsSimSteps with InterestSimSteps
+  with GiftAidSimSteps with EmploymentsSimSteps with StudentLoansSimSteps with CrystallisationSimSteps with NRSSimSteps {
 
-  setup("individual-dividends", "Individual Dividends Journey") withRequests(
-    getLoginPage,
-    postIndividualLoginPage("AA111111A", "1234567890"),
-    getStartPage(taxYear),
-    getOverviewPage(taxYear),
-    getUKDividendsStatusPage,
-    postUKDividendsStatusPage,
-    getUKDividendsAmountPage,
-    postUKDividendsAmountPage,
-    getOtherDividendsStatusPage,
-    postOtherDividendsStatusPage,
-    getOtherDividendsAmountPage,
-    postOtherDividendsAmountPage,
-    getDividendsCheckYourAnswersPage,
-    postDividendsCheckYourAnswersPage
-  )
+  //Dividends
+  individualDividends(id = "individual-dividends", description = "Individual Dividends Journey")
+  agentDividends(id = "agent-dividends", description = "Agent Dividends Journey")
 
-  setup("agent-dividends", "Agent Dividends Journey") withRequests(
-    getLoginPage,
-    postAgentLoginPage("AA111112A", "1234567890"),
-    getInsertAdditionalParametersEndPoint("AA111112A", "1234567890"),
-    getStartPage(taxYear),
-    getOverviewPage(taxYear),
-    getUKDividendsStatusPage,
-    postUKDividendsStatusPage,
-    getUKDividendsAmountPage,
-    postUKDividendsAmountPage,
-    getOtherDividendsStatusPage,
-    postOtherDividendsStatusPage,
-    getOtherDividendsAmountPage,
-    postOtherDividendsAmountPage,
-    getDividendsCheckYourAnswersPage,
-    postDividendsCheckYourAnswersPage
-  )
+  //Interest
+  interestIndividual(id = "individual-interest", description = "Individual Interest Journey")
+  interestAgent(id = "agent-interest", description = "Agent Interest Journey")
 
-  setup("individual-interest", "Individual Interest Journey") withRequests(
-    getLoginPage,
-    postIndividualLoginPage("AA111111A", "1234567890"),
-    getStartPage(taxYear),
-    getOverviewPage(taxYear),
-    getUntaxedUKInterestStatusPage,
-    postUntaxedUKInterestStatusPage,
-    getUntaxedChooseAccountRedirect,
-    getUntaxedUKInterestDetailsPage,
-    postUntaxedUKInterestDetailsPage(accountName = "Tesco Bank"),
-    getUntaxedUKInterestSummaryPage,
-    postUntaxedUKInterestSummaryPage(addAnotherAccount = true),
-    getUntaxedUKInterestDetailsPage,
-    postUntaxedUKInterestDetailsPage(accountName = "Halifax"),
-    getUntaxedUKInterestSummaryPage,
-    postUntaxedUKInterestSummaryPage(addAnotherAccount = true),
-    getUntaxedUKInterestDetailsPage,
-    postUntaxedUKInterestDetailsPage(accountName = "Santander"),
-    getUntaxedUKInterestSummaryPage,
-    postUntaxedUKInterestSummaryPage(addAnotherAccount = false),
-    getTaxedUKInterestStatusPage,
-    postTaxedUKInterestStatusPage,
-    getTaxedChooseAccountPage,
-    postTaxedChooseAccountPage,
-    getTaxedUKInterestDetailsPage,
-    postTaxedUKInterestDetailsPage(accountName = "Tesco Bank"),
-    getTaxedUKInterestSummaryPage,
-    postTaxedUKInterestSummaryPage(addAnotherAccount = true),
-    getTaxedUKInterestDetailsPage,
-    postTaxedUKInterestDetailsPage(accountName = "Halifax"),
-    getTaxedUKInterestSummaryPage,
-    postTaxedUKInterestSummaryPage(addAnotherAccount = true),
-    getTaxedUKInterestDetailsPage,
-    postTaxedUKInterestDetailsPage(accountName = "Santander"),
-    getTaxedUKInterestSummaryPage,
-    postTaxedUKInterestSummaryPage(addAnotherAccount = false),
-    getInterestCheckYourAnswersPage,
-    postInterestCheckYourAnswersPage
-  )
+  //GiftAid
+  giftAidIndividual(id ="individual-gift-aid", description = "Individual Gift Aid Journey")
+  giftAidAgent(id ="agent-gift-aid", description = "Agent Gift Aid Journey")
 
-  setup("agent-interest", "Agent Interest Journey") withRequests(
-    getLoginPage,
-    postAgentLoginPage("AA111112A", "1234567890"),
-    getInsertAdditionalParametersEndPoint("AA111112A", "1234567890"),
-    getStartPage(taxYear),
-    getOverviewPage(taxYear),
-    getUntaxedUKInterestStatusPage,
-    postUntaxedUKInterestStatusPage,
-    getUntaxedChooseAccountRedirect,
-    getUntaxedUKInterestDetailsPage,
-    postUntaxedUKInterestDetailsPage("Tesco Bank"),
-    getUntaxedUKInterestSummaryPage,
-    postUntaxedUKInterestSummaryPage(addAnotherAccount = true),
-    getUntaxedUKInterestDetailsPage,
-    postUntaxedUKInterestDetailsPage("Halifax"),
-    getUntaxedUKInterestSummaryPage,
-    postUntaxedUKInterestSummaryPage(addAnotherAccount = true),
-    getUntaxedUKInterestDetailsPage,
-    postUntaxedUKInterestDetailsPage("Santander"),
-    getUntaxedUKInterestSummaryPage,
-    postUntaxedUKInterestSummaryPage(addAnotherAccount = false),
-    getTaxedUKInterestStatusPage,
-    postTaxedUKInterestStatusPage,
-    getTaxedChooseAccountPage,
-    postTaxedChooseAccountPage,
-    getTaxedUKInterestDetailsPage,
-    postTaxedUKInterestDetailsPage("Tesco Bank"),
-    getTaxedUKInterestSummaryPage,
-    postTaxedUKInterestSummaryPage(addAnotherAccount = true),
-    getTaxedUKInterestDetailsPage,
-    postTaxedUKInterestDetailsPage("Halifax"),
-    getTaxedUKInterestSummaryPage,
-    postTaxedUKInterestSummaryPage(addAnotherAccount = true),
-    getTaxedUKInterestDetailsPage,
-    postTaxedUKInterestDetailsPage("Santander"),
-    getTaxedUKInterestSummaryPage,
-    postTaxedUKInterestSummaryPage(addAnotherAccount = false),
-    getInterestCheckYourAnswersPage,
-    postInterestCheckYourAnswersPage
-  )
+  //Employment
+  employmentIndividual(id ="individual-employment", description = "Individual Single Employment Journey")
+  employmentAgent(id ="agent-employment", description = "Agent Multiple Employments Journey")
+  addEmploymentIndividual(id ="individual-add-employment", description = "Individual Add Employment-No Employments")
+  addEmploymentAgent(id ="agent-add-employment", description = "Agent Add Employment - Prior Employments")
+  removeEmploymentIndividual(id ="individual-remove-employment", description = "Individual Remove Employment - Single Employment")
+  removeEmploymentAgent(id ="agent-remove-employment", description = "Agent Remove Employment - Multiple Employments")
+  employmentDetailsIndividual(id ="individual-employment-details", description = "Individual Multiple Employments - Check Employment Details EOY")
+  employmentBenefitsIndividual(id ="individual-add-benefits", description = "Individual Multiple Employments - Adding benefits full flow")
+  employmentExpensesIndividual(id ="individual-add-expenses", description = "Individual Expenses Journey - Single Employment")
+  employmentExpensesAgent(id ="agent-add-expenses", description = "Agent Expenses Journey - Multiple Employments")
+  addNewEmploymentIndividual(id ="individual-add-new-employment", description = "Individual Add New Employment - Prior Employments")
+  addNewEmploymentAgent(id ="agent-add-new-employment", description = "Agent Add New Employment - Prior Employments")
 
-  setup("individual-employment", "Individual Single Employment Journey") withRequests(
-    getLoginPage,
-    postIndividualLoginPage("AA133742A", "1234567890"),
-    getStartPage(taxYear),
-    getOverviewPage(taxYear),
-    getEmploymentSummaryPage(taxYear),
-    getEmploymentDetailsPage(taxYear, "00000000-0000-1000-8000-000000000000"),
-    getEmploymentBenefitsPage(taxYear, "00000000-0000-1000-8000-000000000000"),
-    getEmploymentExpensesPage(taxYear)
-  )
+  //StudentLoans
+  studentLoansIndividual(id ="individual-student-loans", description = "Individual Student Loans Journey")
+  studentLoansAgent(id ="agent-student-loans", description = "Agent Student Loans Journey")
 
-  setup("agent-employment", "Agent Multiple Employments Journey") withRequests(
-    getLoginPage,
-    postAgentLoginPage("BB444444A", "1234567890"),
-    getInsertAdditionalParametersEndPoint("BB444444A", "1234567890"),
-    getStartPage(taxYear),
-    getOverviewPage(taxYear),
-    getEmploymentSummaryPage(taxYear),
-    getEmployerDetailsAndBenefitsPage(taxYear, employmentId = "00000000-5555-0000-0000-000000000001"),
-    getEmploymentDetailsPage(taxYear, employmentId = "00000000-5555-0000-0000-000000000001"),
-    getEmploymentBenefitsPage(taxYear, employmentId = "00000000-5555-0000-0000-000000000001"),
-    getEmploymentExpensesPage(taxYear)
-  )
+  //Crystallisation
+  crystallisationIndividual(id ="individual-crystallisation", description = "Individual Crystallisation Journey")
+  crystallisationAgent(id ="agent-crystallisation", description = "Agent Crystallisation Journey")
 
-  setup("individual-add-employment", "Individual Add Employment-No Employments") withRequests(
-    getLoginPage,
-    postIndividualLoginPage("AA123456A", "1234567890"),
-    getStartPage(taxYearEOY),
-    getEOYOverviewPage(taxYearEOY),
-    getAddEmploymentPage(taxYearEOY),
-    postAddEmploymentPage(taxYearEOY),
-    getEmployerNamePage(taxYearEOY),
-    postEmployerNamePage(taxYearEOY),
-    getPayeRefPage(taxYearEOY),
-    postPayeRefPage(taxYearEOY),
-    getEmploymentStartDatePage(taxYearEOY),
-    postEmploymentStartDate(taxYearEOY),
-    getAreYouStillWorkingAtEmployerPage(taxYearEOY),
-    postAreYouStillWorkingAtEmployerPage(taxYearEOY),
-    getWhenDidYouLeaveYourEmployerPage(taxYearEOY),
-    postWhenDidYouLeaveYourEmployerPage(taxYearEOY),
-    getPayrollIdPage(taxYearEOY),
-    postPayrollIdPage(taxYearEOY),
-    getPayEmploymentPage(taxYearEOY),
-    postPayEmploymentPage(taxYearEOY),
-    getUkTaxEmploymentPage(taxYearEOY),
-    postUkTaxEmploymentPage(taxYearEOY)
-  )
-
-  setup("agent-add-employment", "Agent Add Employment - Prior Employments") withRequests(
-    getLoginPage,
-    postAgentLoginPage("BB444444A", "1234567890"),
-    getInsertAdditionalParametersEndPoint("BB444444A", "1234567890"),
-    getStartPage(taxYearEOY),
-    getEOYOverviewPage(taxYearEOY),
-    getEmploymentSummaryPage(taxYearEOY),
-    getEmployerNamePage(taxYearEOY),
-    postEmployerNamePage(taxYearEOY),
-    getPayeRefPage(taxYearEOY),
-    postPayeRefPage(taxYearEOY),
-    getEmploymentStartDatePage(taxYearEOY),
-    postEmploymentStartDate(taxYearEOY),
-    getAreYouStillWorkingAtEmployerPage(taxYearEOY),
-    postAreYouStillWorkingAtEmployerPage(taxYearEOY),
-    getWhenDidYouLeaveYourEmployerPage(taxYearEOY),
-    postWhenDidYouLeaveYourEmployerPage(taxYearEOY),
-    getPayrollIdPage(taxYearEOY),
-    postPayrollIdPage(taxYearEOY),
-    getPayEmploymentPage(taxYearEOY),
-    postPayEmploymentPage(taxYearEOY),
-    getUkTaxEmploymentPage(taxYearEOY),
-    postUkTaxEmploymentPage(taxYearEOY)
-  )
-
-  setup("individual-remove-employment", "Individual Remove Employment - Single Employment") withRequests(
-    getLoginPage,
-    postIndividualLoginPage("AA133742A", "1234567890"),
-    getStartPage(taxYearEOY),
-    getEOYOverviewPage(taxYearEOY),
-    getEmploymentSummaryPage(taxYearEOY),
-    getRemoveEmploymentPage(taxYearEOY, "00000000-0000-1000-8000-000000000004"),
-    postRemoveEmploymentPage(taxYearEOY, "00000000-0000-1000-8000-000000000004"),
-    getEmploymentSummaryPage(taxYearEOY) //TODO - change to EOY overview page once caching problem is resolved
-  )
-
-  setup("agent-remove-employment", "Agent Remove Employment - Multiple Employments") withRequests(
-    getLoginPage,
-    postAgentLoginPage("BB444444A", "1234567890"),
-    getInsertAdditionalParametersEndPoint("BB444444A", "1234567890"),
-    getStartPage(taxYearEOY),
-    getEOYOverviewPage(taxYearEOY),
-    getEmploymentSummaryPage(taxYearEOY),
-    getRemoveEmploymentPage(taxYearEOY, "00000000-5555-0000-0000-000000000001"),
-    postRemoveEmploymentPage(taxYearEOY, "00000000-5555-0000-0000-000000000001"),
-    getEmploymentSummaryPage(taxYearEOY)
-  )
-
-  setup("individual-employment-details", "Individual Multiple Employments - Check Employment Details EOY") withRequests(
-    getLoginPage,
-    postIndividualLoginPage("BB444444A", "1234567890"),
-    getStartPage(taxYearEOY),
-    getEOYOverviewPage(taxYearEOY),
-    getEmploymentSummaryPage(taxYearEOY),
-    getEmployerDetailsAndBenefitsPage(taxYearEOY, "00000000-5555-5555-0000-000000000002"),
-    getEmploymentDetailsPage(taxYearEOY, "00000000-5555-5555-0000-000000000002"),
-    postEmploymentDetailsPage(taxYearEOY, "00000000-5555-5555-0000-000000000002"),
-    getEmploymentSummaryPage(taxYearEOY)
-  )
-
-  setup("individual-add-benefits", "Individual Multiple Employments - Adding benefits full flow") withRequests(
-    getLoginPage,
-    postIndividualLoginPage("BB444444A", "1234567890"),
-    getStartPage(taxYearEOY),
-    getEOYOverviewPage(taxYearEOY),
-    getEmploymentSummaryPage(taxYearEOY),
-    getEmployerDetailsAndBenefitsPage(taxYearEOY, "00000000-5555-0000-0000-000000000002"),
-    getEmploymentBenefitsWithoutBenefitsPage(taxYearEOY, "00000000-5555-0000-0000-000000000002"),
-    getReceivedBenefitsPage,
-    postReceivedBenefitsPage,
-    getCarVanFuelBenefitsPage,
-    postCarVanFuelBenefitsPage,
-    getCompanyCarBenefitsPage,
-    postCompanyCarBenefitsPage,
-    getCompanyCarAmountBenefitsPage,
-    postCompanyCarAmountBenefitsPage,
-    getCompanyCarFuelBenefitsPage,
-    postCompanyCarFuelBenefitsPage,
-    getCompanyCarFuelAmountBenefitsPage,
-    postCompanyCarFuelAmountBenefitsPage,
-    getCompanyVanBenefitsPage,
-    postCompanyVanBenefitsPage,
-    getCompanyVanAmountBenefitsPage,
-    postCompanyVanAmountBenefitsPage,
-    getCompanyVanFuelBenefitsPage,
-    postCompanyVanFuelBenefitsPage,
-    getCompanyVanFuelAmountBenefitsPage,
-    postCompanyVanFuelAmountBenefitsPage,
-    getMileageBenefitsPage,
-    postMileageBenefitsPage,
-    getMileageBenefitsAmountPage,
-    postMileageBenefitsAmountPage,
-    getAccommodationRelocationBenefitsPage,
-    postAccommodationRelocationBenefitsPage,
-    getAccommodationBenefitsPage,
-    postAccommodationBenefitsPage,
-    getAccommodationAmountBenefitsPage,
-    postAccommodationAmountBenefitsPage,
-    getQualifyingRelocationBenefitsPage,
-    postQualifyingRelocationBenefitsPage,
-    getQualifyingRelocationAmountBenefitsPage,
-    postQualifyingRelocationAmountBenefitsPage,
-    getNonQualifyingRelocationBenefitsPage,
-    postNonQualifyingRelocationBenefitsPage,
-    getNonQualifyingRelocationAmountBenefitsPage,
-    postNonQualifyingRelocationAmountBenefitsPage,
-    getTravelOrEntertainmentBenefitsPage,
-    postTravelOrEntertainmentBenefitsPage,
-    getTravelOrSubsistenceBenefitsPage,
-    postTravelOrSubsistenceBenefitsPage,
-    getTravelOrSubsistenceAmountBenefitsPage,
-    postTravelOrSubsistenceAmountBenefitsPage,
-    getIncidentalOvernightCostsBenefitsPage,
-    postIncidentalOvernightCostsBenefitsPage,
-    getIncidentalOvernightCostsAmountBenefitsPage,
-    postIncidentalOvernightCostsAmountBenefitsPage,
-    getEntertainmentBenefitsPage,
-    postEntertainmentBenefitsPage,
-    getEntertainmentAmountBenefitsPage,
-    postEntertainmentAmountBenefitsPage,
-    getUtilitiesBenefitsPage,
-    postUtilitiesBenefitsPage,
-    getTelephoneBenefitsPage,
-    postTelephoneBenefitsPage,
-    getTelephoneAmountBenefitsPage,
-    postTelephoneAmountBenefitsPage,
-    getEmployerServicesBenefitsPage,
-    postEmployerServicesBenefitsPage,
-    getEmployerServicesAmountBenefitsPage,
-    postEmployerServicesAmountBenefitsPage,
-    getProfessionalFeesBenefitsPage,
-    postProfessionalFeesBenefitsPage,
-    getProfessionalFeesAmountBenefitsPage,
-    postProfessionalFeesAmountBenefitsPage,
-    getOtherServicesBenefitsPage,
-    postOtherServicesBenefitsPage,
-    getOtherServicesAmountBenefitsPage,
-    postOtherServicesAmountBenefitsPage,
-    getMedicalDentalChildcareBenefitsPage,
-    postMedicalDentalChildcareBenefitsPage,
-    getMedicalDentalBenefitsPage,
-    postMedicalDentalBenefitsPage,
-    getMedicalDentalAmountBenefitsPage,
-    postMedicalDentalAmountBenefitsPage,
-    getChildcareBenefitsPage,
-    postChildcareBenefitsPage,
-    getChildcareAmountBenefitsPage,
-    postChildcareAmountBenefitsPage,
-    getEducationalServicesBenefitsPage,
-    postEducationalServicesBenefitsPage,
-    getEducationalServicesAmountBenefitsPage,
-    postEducationalServicesAmountBenefitsPage,
-    getBeneficialLoansBenefitsPage,
-    postBeneficialLoansBenefitsPage,
-    getBeneficialLoansAmountBenefitsPage,
-    postBeneficialLoansAmountBenefitsPage,
-    getIncomeTaxOrIncurredCostsBenefitsPage,
-    postIncomeTaxOrIncurredCostsBenefitsPage,
-    getIncomeTaxBenefitsPage,
-    postIncomeTaxBenefitsPage,
-    getIncomeTaxAmountBenefitsPage,
-    postIncomeTaxAmountBenefitsPage,
-    getIncurredCostsBenefitsPage,
-    postIncurredCostsBenefitsPage,
-    getIncurredCostsAmountBenefitsPage,
-    postIncurredCostsAmountBenefitsPage,
-    getReimbursedCostsVouchersNonCashBenefitsPage,
-    postReimbursedCostsVouchersNonCashBenefitsPage,
-    getNonTaxableCostsBenefitsPage,
-    postNonTaxableCostsBenefitsPage,
-    getNonTaxableCostsAmountBenefitsPage,
-    postNonTaxableCostsAmountBenefitsPage,
-    getTaxableCostsBenefitsPage,
-    postTaxableCostsBenefitsPage,
-    getTaxableCostsAmountBenefitsPage,
-    postTaxableCostsAmountBenefitsPage,
-    getVouchersOrCreditCardsBenefitsPage,
-    postVouchersOrCreditCardsBenefitsPage,
-    getVouchersOrCreditCardsAmountBenefitsPage,
-    postVouchersOrCreditCardsAmountBenefitsPage,
-    getNonCashBenefitsPage,
-    postNonCashBenefitsPage,
-    getNonCashAmountBenefitsPage,
-    postNonCashAmountBenefitsPage,
-    getOtherBenefitsPage,
-    postOtherBenefitsPage,
-    getOtherBenefitsAmountPage,
-    postOtherBenefitsAmountPage,
-    getAssetsOrAssetsTransferBenefitsPage,
-    postAssetsOrAssetsTransferBenefitsPage,
-    getAssetsBenefitsPage,
-    postAssetsBenefitsPage,
-    getAssetsBenefitsAmountPage,
-    postAssetsBenefitsAmountPage,
-    getAssetsTransferBenefitsPage,
-    postAssetsTransferBenefitsPage,
-    getAssetsTransferBenefitsAmountPage,
-    postAssetsTransferBenefitsAmountPage,
-    getEmploymentBenefitsPage(taxYearEOY, "00000000-5555-0000-0000-000000000002")
-  )
-
-  setup("individual-add-expenses", "Individual Expenses Journey - Single Employment") withRequests(
-    getLoginPage,
-    postIndividualLoginPage("AA133742A", "1234567890"),
-    getStartPage(taxYearEOY),
-    getEOYOverviewPage(taxYearEOY),
-    getEmploymentSummaryPage(taxYearEOY),
-    getEmployerDetailsAndBenefitsPage(taxYearEOY, "00000000-0000-1000-8000-000000000004"),
-    getEmploymentExpensesPage(taxYearEOY),
-    getClaimingExpensesPage,
-    postClaimingExpensesPage,
-    getTravelAndOvernightExpensesPage,
-    postTravelAndOvernightExpensesPage,
-    getTravelAndOvernightExpensesAmountPage,
-    postTravelAndOvernightExpensesAmountPage,
-    getUniformsOrToolsExpensesPage,
-    postUniformsOrToolsExpensesPage,
-    getUniformsOrToolsExpensesAmountPage,
-    postUniformsOrToolsExpensesAmountPage,
-    getProfessionalSubscriptionsExpensesPage,
-    postProfessionalSubscriptionsExpensesPage,
-    getProfessionalSubscriptionsExpensesAmountPage,
-    postProfessionalSubscriptionsExpensesAmountPage,
-    getOtherEquipmentExpensesPage,
-    postOtherEquipmentExpensesPage,
-    getOtherEquipmentExpensesAmountPage,
-    postOtherEquipmentExpensesAmountPage
-  )
-
-  setup("agent-add-expenses", "Agent Expenses Journey - Multiple Employments") withRequests(
-    getLoginPage,
-    postAgentLoginPage("BB444444A", "1234567890"),
-    getInsertAdditionalParametersEndPoint("BB444444A", "1234567890"),
-    getStartPage(taxYearEOY),
-    getEOYOverviewPage(taxYearEOY),
-    getEmploymentSummaryPage(taxYearEOY),
-    getEmploymentExpensesPage(taxYearEOY),
-    getClaimingExpensesPage,
-    postClaimingExpensesPage,
-    getTravelAndOvernightExpensesPage,
-    postTravelAndOvernightExpensesPage,
-    getTravelAndOvernightExpensesAmountPage,
-    postTravelAndOvernightExpensesAmountPage,
-    getUniformsOrToolsExpensesPage,
-    postUniformsOrToolsExpensesPage,
-    getUniformsOrToolsExpensesAmountPage,
-    postUniformsOrToolsExpensesAmountPage,
-    getProfessionalSubscriptionsExpensesPage,
-    postProfessionalSubscriptionsExpensesPage,
-    getProfessionalSubscriptionsExpensesAmountPage,
-    postProfessionalSubscriptionsExpensesAmountPage,
-    getOtherEquipmentExpensesPage,
-    postOtherEquipmentExpensesPage,
-    getOtherEquipmentExpensesAmountPage,
-    postOtherEquipmentExpensesAmountPage
-  )
-
-  setup("individual-gift-aid", "Individual Gift Aid Journey") withRequests(
-    getLoginPage,
-    postIndividualLoginPage("AA111112A", "1234567890"),
-    getStartPage(taxYear),
-    getOverviewPage(taxYear),
-    getGiftAidStatusPage,
-    postGiftAidStatusPage,
-    getGiftAidDetailsPage,
-    postGiftAidDetailsPage,
-    getOneOffDonationStatusPage,
-    postOneOffDonationStatusPage,
-    getOneOffDonationDetailsPage,
-    postOneOffDonationDetailsPage,
-    getOverseasCharityDonationStatusPage,
-    postOverseasCharityDonationStatusPage,
-    getOverseasCharityDonationDetailsPage,
-    postOverseasCharityDonationDetailsPage,
-    getNameOfOverseasCharityDetailsPage,
-    postNameOfOverseasCharityDetailsPage(charityName = "French Charity"),
-    getOverseasCharitiesDonatedToSummaryPage,
-    postOverseasCharitiesDonatedToSummaryPage(addAnotherCharity = true),
-    getNameOfOverseasCharityDetailsPage,
-    postNameOfOverseasCharityDetailsPage(charityName = "Spanish Charity"),
-    getOverseasCharitiesDonatedToSummaryPage,
-    postOverseasCharitiesDonatedToSummaryPage(addAnotherCharity = true),
-    getNameOfOverseasCharityDetailsPage,
-    postNameOfOverseasCharityDetailsPage(charityName = "Italian Charity"),
-    getOverseasCharitiesDonatedToSummaryPage,
-    postOverseasCharitiesDonatedToSummaryPage(addAnotherCharity = false),
-    getAddDonationsToLastTaxYearStatusPage,
-    postAddDonationsToLastTaxYearStatusPage,
-    getAddDonationsToLastTaxYearDetailsPage,
-    postAddDonationsToLastTaxYearDetailsPage,
-    getAddDonationsCurrentTaxYearStatusPage,
-    postAddDonationsCurrentTaxYearStatusPage,
-    getAddDonationsCurrentTaxYearDetailsPage,
-    postAddDonationsCurrentTaxYearDetailsPage,
-    getDonationOfSharesOrSecuritiesStatusPage,
-    postDonationOfSharesOrSecuritiesStatusPage,
-    getValueOfSharesOrSecuritiesDetailsPage,
-    postValueOfSharesOrSecuritiesDetailsPage,
-    getDonationOfLandOrPropertyStatusPage,
-    postDonationOfLandOrPropertyStatusPage,
-    getValueOfLandOrPropertyDetailsPage,
-    postValueOfLandOrPropertyDetailsPage,
-    getDonationOfSharesSecuritiesLandPropertyOverseasStatusPage,
-    postDonationOfSharesSecuritiesLandPropertyOverseasStatusPage,
-    getValueOfSharesSecuritiesLandPropertyOverseasDetailsPage,
-    postValueOfSharesSecuritiesLandPropertyOverseasDetailsPage,
-    getNameOfOverseasCharitySharesSecuritiesLandPropertyDetailsPage,
-    postNameOfOverseasCharitySharesSecuritiesLandPropertyDetailsPage(charityName = "Angolan Charity"),
-    getOverseasCharitiesSharesSecuritiesLandPropertyDonatedToSummaryPage,
-    postOverseasCharitiesSharesSecuritiesLandPropertyDonatedToSummaryPage(addAnotherCharity = true),
-    getNameOfOverseasCharitySharesSecuritiesLandPropertyDetailsPage,
-    postNameOfOverseasCharitySharesSecuritiesLandPropertyDetailsPage(charityName = "Nepalese Charity"),
-    getOverseasCharitiesSharesSecuritiesLandPropertyDonatedToSummaryPage,
-    postOverseasCharitiesSharesSecuritiesLandPropertyDonatedToSummaryPage(addAnotherCharity = true),
-    getNameOfOverseasCharitySharesSecuritiesLandPropertyDetailsPage,
-    postNameOfOverseasCharitySharesSecuritiesLandPropertyDetailsPage(charityName = "Antarctican Charity"),
-    getOverseasCharitiesSharesSecuritiesLandPropertyDonatedToSummaryPage,
-    postOverseasCharitiesSharesSecuritiesLandPropertyDonatedToSummaryPage(addAnotherCharity = false),
-    getGiftAidCheckYourAnswersPage,
-    postGiftAidCheckYourAnswersPage
-  )
-
-  setup("agent-gift-aid", "Agent Gift Aid Journey") withRequests(
-    getLoginPage,
-    postAgentLoginPage("AA111112A", "1234567890"),
-    getInsertAdditionalParametersEndPoint("AA111112A", "1234567890"),
-    getStartPage(taxYear),
-    getOverviewPage(taxYear),
-    getGiftAidStatusPage,
-    postGiftAidStatusPage,
-    getGiftAidDetailsPage,
-    postGiftAidDetailsPage,
-    getOneOffDonationStatusPage,
-    postOneOffDonationStatusPage,
-    getOneOffDonationDetailsPage,
-    postOneOffDonationDetailsPage,
-    getOverseasCharityDonationStatusPage,
-    postOverseasCharityDonationStatusPage,
-    getOverseasCharityDonationDetailsPage,
-    postOverseasCharityDonationDetailsPage,
-    getNameOfOverseasCharityDetailsPage,
-    postNameOfOverseasCharityDetailsPage(charityName = "French Charity"),
-    getOverseasCharitiesDonatedToSummaryPage,
-    postOverseasCharitiesDonatedToSummaryPage(addAnotherCharity = true),
-    getNameOfOverseasCharityDetailsPage,
-    postNameOfOverseasCharityDetailsPage(charityName = "Spanish Charity"),
-    getOverseasCharitiesDonatedToSummaryPage,
-    postOverseasCharitiesDonatedToSummaryPage(addAnotherCharity = true),
-    getNameOfOverseasCharityDetailsPage,
-    postNameOfOverseasCharityDetailsPage(charityName = "Italian Charity"),
-    getOverseasCharitiesDonatedToSummaryPage,
-    postOverseasCharitiesDonatedToSummaryPage(addAnotherCharity = false),
-    getAddDonationsToLastTaxYearStatusPage,
-    postAddDonationsToLastTaxYearStatusPage,
-    getAddDonationsToLastTaxYearDetailsPage,
-    postAddDonationsToLastTaxYearDetailsPage,
-    getAddDonationsCurrentTaxYearStatusPage,
-    postAddDonationsCurrentTaxYearStatusPage,
-    getAddDonationsCurrentTaxYearDetailsPage,
-    postAddDonationsCurrentTaxYearDetailsPage,
-    getDonationOfSharesOrSecuritiesStatusPage,
-    postDonationOfSharesOrSecuritiesStatusPage,
-    getValueOfSharesOrSecuritiesDetailsPage,
-    postValueOfSharesOrSecuritiesDetailsPage,
-    getDonationOfLandOrPropertyStatusPage,
-    postDonationOfLandOrPropertyStatusPage,
-    getValueOfLandOrPropertyDetailsPage,
-    postValueOfLandOrPropertyDetailsPage,
-    getDonationOfSharesSecuritiesLandPropertyOverseasStatusPage,
-    postDonationOfSharesSecuritiesLandPropertyOverseasStatusPage,
-    getValueOfSharesSecuritiesLandPropertyOverseasDetailsPage,
-    postValueOfSharesSecuritiesLandPropertyOverseasDetailsPage,
-    getNameOfOverseasCharitySharesSecuritiesLandPropertyDetailsPage,
-    postNameOfOverseasCharitySharesSecuritiesLandPropertyDetailsPage(charityName = "Angolan Charity"),
-    getOverseasCharitiesSharesSecuritiesLandPropertyDonatedToSummaryPage,
-    postOverseasCharitiesSharesSecuritiesLandPropertyDonatedToSummaryPage(addAnotherCharity = true),
-    getNameOfOverseasCharitySharesSecuritiesLandPropertyDetailsPage,
-    postNameOfOverseasCharitySharesSecuritiesLandPropertyDetailsPage(charityName = "Nepalese Charity"),
-    getOverseasCharitiesSharesSecuritiesLandPropertyDonatedToSummaryPage,
-    postOverseasCharitiesSharesSecuritiesLandPropertyDonatedToSummaryPage(addAnotherCharity = true),
-    getNameOfOverseasCharitySharesSecuritiesLandPropertyDetailsPage,
-    postNameOfOverseasCharitySharesSecuritiesLandPropertyDetailsPage(charityName = "Antarctican Charity"),
-    getOverseasCharitiesSharesSecuritiesLandPropertyDonatedToSummaryPage,
-    postOverseasCharitiesSharesSecuritiesLandPropertyDonatedToSummaryPage(addAnotherCharity = false),
-    getGiftAidCheckYourAnswersPage,
-    postGiftAidCheckYourAnswersPage
-  )
-
-  setup("individual-crystallisation", "Individual Crystallisation Journey") withRequests(
-    getLoginPage,
-    postIndividualLoginPage("AA888888A", "XAIT00000888888"),
-    getStartPage(taxYearEOY),
-    getViewChangeTestHeadersPage,
-    postViewChangeTestHeadersPage,
-    getEOYOverviewPage(taxYearEOY),
-    postEOYOverviewPage(taxYearEOY),
-    getFinalTaxCalculationEndpoint(taxYearEOY),
-    getFinalTaxCalculationPage(taxYearEOY),
-    postFinalTaxCalculationPage(taxYearEOY),
-    getDeclarationPage(taxYearEOY),
-    postDeclarationPage(taxYearEOY),
-    getReceivedTaxCalculationPage(taxYearEOY)
-  )
-
-  setup("agent-crystallisation", "Agent Crystallisation Journey") withRequests(
-    getLoginPage,
-    postAgentLoginPage("AA888888A", "XAIT00000888888"),
-    getInsertAdditionalParametersEndPoint("AA888888A", "XAIT00000888888"),
-    getStartPage(taxYearEOY),
-    getEOYOverviewPage(taxYearEOY),
-    postEOYOverviewPage(taxYearEOY),
-    getCitizenDetailsStubPage,
-    postCitizenDetailsStubPage,
-    getClientUTRPage,
-    postClientUTRPage,
-    getConfirmClientUTRPage,
-    postConfirmClientUTRPage,
-    getViewChangeTestHeadersPage,
-    postViewChangeTestHeadersPage,
-    getFinalTaxCalculationAgentEndpoint(taxYearEOY),
-    getFinalTaxCalculationAgentPage(taxYearEOY),
-    postFinalTaxCalculationAgentPage(taxYearEOY),
-    getDeclarationPage(taxYearEOY),
-    postDeclarationPage(taxYearEOY),
-    getReceivedTaxCalculationPage(taxYearEOY)
-  )
-
-  setup("nrs-proxy-crystallisation", "NRS Crystallisation Downstream Request") withRequests(
-    insertAuthRecordInd("AA888888A", "XAIT00000888888"),
-    postNrsProxyStatement("AA888888A", "itsa-crystallisation", "XAIT00000888888")
-
-  )
-
-  setup("individual-add-new-employment", "Individual Add New Employment - Prior Employments") withRequests(
-    getLoginPage,
-    postIndividualLoginPage("BB444444A", "XAIT00000888888"),
-    getStartPage(taxYearEOY),
-    getEOYOverviewPage(taxYearEOY),
-    getEmploymentSummaryPage(taxYearEOY),
-    getEmployerDetailsAndBenefitsPage(taxYearEOY, "00000000-5555-0000-0000-000000000002"),
-    getEmploymentDetailsPage(taxYearEOY, "00000000-5555-0000-0000-000000000002"),
-    postEmploymentDetailsPage(taxYearEOY, "000000000-5555-0000-0000-000000000002"),
-    getReceivedBenefitsPage,
-    postReceivedBenefitsPage,
-    getEmploymentBenefitsPage(taxYearEOY, "00000000-5555-0000-0000-000000000002"),
-    getClaimingExpensesPage,
-    postClaimingExpensesPage,
-    getEmploymentSummaryPage(taxYearEOY)
-
-  )
-
-  setup("agent-add-new-employment", "Agent Add New Employment - Prior Employments") withRequests(
-    getLoginPage,
-    postAgentLoginPage("BB444444A", "1234567890"),
-    getInsertAdditionalParametersEndPoint("BB444444A", "1234567890"),
-    getStartPage(taxYearEOY),
-    getEOYOverviewPage(taxYearEOY),
-    getEmploymentSummaryPage(taxYearEOY),
-    getEmployerDetailsAndBenefitsPage(taxYearEOY, "00000000-5555-0000-0000-000000000002"),
-    getEmploymentDetailsPage(taxYearEOY, "00000000-5555-0000-0000-000000000002"),
-    postEmploymentDetailsPage(taxYearEOY, "00000000-5555-0000-0000-000000000002"),
-    getReceivedBenefitsPage,
-    postReceivedBenefitsPage,
-    getEmploymentBenefitsPage(taxYearEOY, "00000000-5555-0000-0000-000000000002"),
-    getClaimingExpensesPage,
-    postClaimingExpensesPage,
-    getEmploymentSummaryPage(taxYearEOY)
-  )
+  //NRS
+  nrsJourney(id ="nrs-proxy-crystallisation", description = "NRS Crystallisation Downstream Request")
 
   runSimulation()
 }
