+++ src/test/scala/uk/gov/hmrc/perftests/itsass/requests/CrystallisationRequests.scala
@@ -14,26 +14,17 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.perftests.itsass
+package uk.gov.hmrc.perftests.itsass.requests
 
 import io.gatling.core.Predef._
 import io.gatling.http.Predef._
 import io.gatling.http.request.builder.HttpRequestBuilder
 import uk.gov.hmrc.performance.conf.ServicesConfiguration
-import uk.gov.hmrc.perftests.itsass.RequestsHelper._
+import uk.gov.hmrc.perftests.itsass.requests.RequestsHelper._
 
 object CrystallisationRequests extends ServicesConfiguration {
 
   //Common
-  def getEOYOverviewPage(taxYearEOY: String): HttpRequestBuilder = http("Get EOY Overview Page")
-    .get(s"$serviceUrl/$taxYearEOY/income-tax-return-overview")
-    .check(saveCsrfToken())
-    .check(status.is(200))
-
-  def postEOYOverviewPage(taxYearEOY: String): HttpRequestBuilder = http("Post EOY Overview Page")
-    .post(s"$serviceUrl/$taxYearEOY/final-calculation")
-    .formParam("""csrfToken""", """${csrfToken}""")
-    .check(status.is(303))
 
   def getViewChangeTestHeadersPage: HttpRequestBuilder = http("Get V&C Test Headers Page")
     .get(s"$viewAndChangeUrl/view/test-only/test-headers")
