+++ src/test/scala/uk/gov/hmrc/perftests/itsass/requests/ExpensesRequests.scala
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.perftests.itsass
+package uk.gov.hmrc.perftests.itsass.requests
 
 import io.gatling.core.Predef._
 import io.gatling.http.Predef._
 import io.gatling.http.request.builder.HttpRequestBuilder
 import uk.gov.hmrc.performance.conf.ServicesConfiguration
-import uk.gov.hmrc.perftests.itsass.RequestsHelper._
+import uk.gov.hmrc.perftests.itsass.requests.RequestsHelper._
 
-object ExpensesRequests extends ServicesConfiguration{
+object ExpensesRequests extends ServicesConfiguration {
 
   def getClaimingExpensesPage: HttpRequestBuilder = http("Get Claim Employment Expenses Question Page")
     .get(s"${employmentsUrl(taxYearEOY)}/expenses/claim-employment-expenses")
