+++ src/test/scala/uk/gov/hmrc/perftests/itsass/simSteps/NRSSimSteps.scala
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2022 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.perftests.itsass.simSteps
+
+import uk.gov.hmrc.performance.simulation.{JourneyPart, PerformanceTestRunner}
+import uk.gov.hmrc.perftests.itsass.requests.AuthLoginAPIRequests.insertAuthRecordInd
+import uk.gov.hmrc.perftests.itsass.requests.NRSProxyRequests.postNrsProxyStatement
+
+trait NRSSimSteps extends PerformanceTestRunner {
+
+  def nrsJourney(id: String, description: String): JourneyPart = setup(id, description) withRequests(
+    insertAuthRecordInd("AA888888A", "XAIT00000888888"),
+    postNrsProxyStatement("AA888888A", "itsa-crystallisation", "XAIT00000888888")
+  )
+
+}
